Solve problems on creating new heuristic, it needs refactoring

+++ src/br/com/representative/clustering/partitioning/KMeans.java
@@ -70,7 +70,7 @@ public class KMeans extends KMethod {
 //                 }
 //                 newGuess[i] = new Point.Double(x/cluster.size(), y/cluster.size());                 
                  
-                 float[] sum = new float[items.get(0).vector().length];
+                 double[] sum = new double[items.get(0).vector().length];
                  Arrays.fill(sum, 0.0f);
                  for( int j = 0; j < cluster.size(); ++j ) {
                      for( int k = 0; k < sum.length; ++k )
